+++ Packages/com.unity.inputsystem/InputSystem/Plugins/XR/WindowsMR.cs
@@ -20,6 +20,9 @@ namespace UnityEngine.InputSystem.XR
 
         protected override void FinishSetup(InputDeviceBuilder builder)
         {
+            if (builder == null)
+                throw new System.ArgumentNullException(nameof(builder));
+
             base.FinishSetup(builder);
 
             trackingState = builder.GetControl<IntegerControl>("trackingState");
@@ -62,6 +65,9 @@ namespace UnityEngine.InputSystem.XR
 
         protected override void FinishSetup(InputDeviceBuilder builder)
         {
+            if (builder == null)
+                throw new System.ArgumentNullException(nameof(builder));
+
             base.FinishSetup(builder);
 
             joystick = builder.GetControl<Vector2Control>("joystick");